Specify default group for damage-animals flag
As of latest version of WG and WGEx the default group for damage-animals flag is all, which causes confusion (as it is incoherent with behavior of other state flags, used to prevent grief (such as enderpearl, lighter or potionsplash)). This commit changes the default RegionGroup for the damage-animals flag to RegionGroup.NON_MEMBERS (others), which is the behavior expected by many (including server administrators) and predicted by other grief-prevention state flag.

An alternative way to solve the issue of confusion with the behavior is to specify the default group under configuration node customflags.damage-animals.defaultgroup.